Fix clang warning.
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@11598 a1ca3aef-8c08-0410-bb20-df032aa958be

+++ cups/adminutil.c
@@ -2165,7 +2165,7 @@ get_cupsd_conf(
     * Read cupsd.conf via a HTTP GET request...
     */
 
-    if ((fd = cupsTempFd(name, namesize)) < 0)
+    if ((fd = cupsTempFd(name, (int)namesize)) < 0)
     {
       *name = '\0';
